refactor: move unique_id tag to creedicationary_extras
