+++ src/CreeDictionary/CreeDictionary/templates/CreeDictionary/components/citations.html
@@ -1,3 +1,7 @@
 {% for source in dictionary_sources %}
   <cite class="cite-dict">{{ source }}</cite>
+  <div id="tooltip:{{ id }}" class="tooltip" role="tooltip">
+    {{ lemma.inflectional_category_linguistic }} ({{ lemma.inflectional_category }})
+    <div class="tooltip__arrow" data-popper-arrow></div>
+  </div>
 {% endfor %}
