+++ src/CreeDictionary/CreeDictionary/templates/CreeDictionary/components/definition__elaboration--verbose.html
@@ -22,7 +22,6 @@
   {% endcomment %}
 
   {% load creedictionary_extras %}
-  {% load inflection_extras %}
 
   {% if ic_plain_english or emoji %}
     <p class="elaboration__row elaboration__wordclass-row">
