+++ src/CreeDictionary/CreeDictionary/templates/CreeDictionary/components/definition__elaboration.html
@@ -28,7 +28,6 @@
   {% endcomment %}
 
   {% load creedictionary_extras %}
-  {% load inflection_extras %}
 
   {# Second line of the header: the elaboration #}
   <div class="definition__elaboration" data-cy="elaboration">
