+++ src/CreeDictionary/CreeDictionary/templates/CreeDictionary/word-detail.html
@@ -14,7 +14,6 @@
   {% endcomment %}
 
   {% load static %}
-  {% load inflection_extras %}
   {% load morphodict_orth %}
 
   <article id="definition" class="definition">
